Add schema.org hours, address, and contact info
These are probably the most important parts of the site for machines to
understand: how do I get to your library, when is it open, and how do I get in
contact with you when I have a problem? All of these things tend to show up on
Google Maps, for example.

Signed-off-by: Dan Scott <dan@coffeecode.net>